refactor: extract config and eval logic into settings.py and eval.py
Extract foundational code from server.py into two new modules:

- settings.py: MathMCPSettings class, validated_tool decorator, static
  constants (MATH_FUNCTIONS_*, DANGEROUS_PATTERNS, ALLOWED_*, TOPIC_KEYWORDS,
  TEMP_CONVERSIONS)
- eval.py: _validate_expression_syntax, safe_eval_expression,
  validate_variable_name, convert_temperature, _classify_expression_*

Keep settings instantiation, MAX_* derivations, evaluate_with_timeout,
requires_matplotlib, and validate_nested_array_groups in server.py to
preserve importlib.reload and mock.patch test semantics.

Re-exports maintain backward compatibility for all test imports.

Part 1 of 4 for #140.
Closes #154
